Python/requirements.txt
phonenumbers
requests
python-dotenv